pyproject.toml
[tool.black]
line-length = 79
target-version = ["py310"]
include = '\.pyi?$'
skip-string-normalization = true
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 79
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ArkDamage"
version = "0.1.0"
description = "明日方舟DPS计算器python版本,基于nonebot2"
authors = ["qwerdvd <2450899274@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.3"
nonebot2 = {extras = ["fastapi"], version = "^2.0.0rc3"}
nonebot-adapter-onebot = ">=2.1.1"
nonebot-plugin-apscheduler = ">=0.1.4"
pydantic = "^1.10.4"
aiohttp = "^3.8.1"
aiofiles = "^0.8.0"

[[tool.poetry.source]]
name = "USTC"
url = "https://pypi.mirrors.ustc.edu.cn/simple"
default = false
secondary = true

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.11.5"
pre-commit = "^2.21.0"
pycln = "^2.1.2"

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
